Move everything OIDC-related to a separate file
